requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
xgboost
joblib
pickle-mixin
lightgbm

# NLP
nltk
transformers
torch

# Image / DL
tensorflow
keras
pillow

# Audio
librosa
soundfile
resampy

# Video
opencv-python

# Utils
matplotlib